requirements-discordpy.txt
discord.py[voice]
python-dotenv
discord-ext-voice-recv
